% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.cancer.signature.R
\name{Get.cancer.signature}
\alias{Get.cancer.signature}
\title{get cancer signature from DEG matirx}
\usage{
Get.cancer.signature(adj.pvalue, log2fc, max.num = FALSE)
}
\arguments{
\item{adj.pvalue}{Set a adjusted p-value to screen for cancer signatures}

\item{log2fc}{Set up a log2fc to filter cancer signatures}

\item{max.num}{Set the upper line value for the number of genes in the cancer signature}
}
\value{
a dataframe with cancer signature information
}
\description{
get cancer signature from DEG matirx
}
