% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.top.signature.R
\name{Get.top.signature}
\alias{Get.top.signature}
\title{Obtain the specified number of genes as cancer signatures used for enrichment analysis}
\usage{
Get.top.signature(top.num = 100)
}
\value{
a matrix with cancer signature information
}
\description{
Obtain the specified number of genes as cancer signatures used for enrichment analysis
}
