% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISEA.R
\name{ISEA.Plot}
\alias{ISEA.Plot}
\title{Visually display the results of ISEA enrichment analysis}
\usage{
ISEA.Plot(sig.mtx, method = "BH")
}
\arguments{
\item{sig.mtx}{A cancer signature matrix after ISEA enrichment analysis}

\item{method}{BH/bonferroni}
}
\value{
Visualization of ISEA enrichment analysis results
}
\description{
Visually display the results of ISEA enrichment analysis
}
