% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISEA.R
\name{ISEA}
\alias{ISEA}
\title{Enrichment analysis of differentially expressed genes based on cancer signature}
\usage{
ISEA(DEG.gene.select, sig.mtx, species = "human")
}
\arguments{
\item{DEG.gene.select}{Differentially expressed genes list}

\item{sig.mtx}{Cancer signature matirx}

\item{species}{Specify which species to perform gene enrichment analysis. The default is human. You can also fill in mouse or zebrafish.}
}
\value{
Matrix containing ISEA result
}
\description{
Enrichment analysis of differentially expressed genes based on cancer signature
}
