% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISEA.R
\name{mouse2human}
\alias{mouse2human}
\title{Convert mouse genes into corresponding human homologous genes}
\usage{
mouse2human(x)
}
\arguments{
\item{x}{A vector containing mouse genes}
}
\value{
Matrix containing human and mouse homologous genes
}
\description{
Convert mouse genes into corresponding human homologous genes
}
