% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.cancer.signature.R
\name{Signature.Plot}
\alias{Signature.Plot}
\title{Visual display of the number of genes included in the cancer signature}
\usage{
Signature.Plot(sig.mtx)
}
\arguments{
\item{sig.mtx}{a matrix containing cancer signatures}
}
\description{
Visual display of the number of genes included in the cancer signature
}
