% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.top.signature.R
\name{Sort_mtx}
\alias{Sort_mtx}
\title{Sort matrix based on corrected p-value and log2 fold change}
\usage{
Sort_mtx(mtx)
}
\arguments{
\item{mtx}{Matrix of differentially expressed genes}
}
\value{
a sorted matrix
}
\description{
Sort matrix based on corrected p-value and log2 fold change
}
