% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISEA.R
\name{zebrafish2human}
\alias{zebrafish2human}
\title{Convert zebrafish genes into corresponding human homologous genes}
\usage{
zebrafish2human(x)
}
\value{
Matrix containing human and zebrafish homologous genes
}
\description{
Convert zebrafish genes into corresponding human homologous genes
}
